+++ lib/arcane_helper_utils.dart
@@ -1,5 +1,6 @@
 library arcane_helper_utils;
 
+export "package:arcane_helper_utils/src/classes/fixed_size_list.dart";
 export "package:arcane_helper_utils/src/extensions/date_time.dart";
 export "package:arcane_helper_utils/src/extensions/dynamic.dart";
 export "package:arcane_helper_utils/src/extensions/jwt.dart";